tracking: eliminate code duplication
